+++ src/Rdd.Infra/Storage/ISaveEventProcessor.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore.ChangeTracking;
+
+namespace Rdd.Infra.Storage
+{
+    public interface ISaveEventProcessor
+    {
+        Task<ISavedEntries> InternalBeforeSaveChangesAsync(ChangeTracker changeTracker);
+        Task InternalAfterSaveChangesAsync(ISavedEntries savedEntries);
+    }
+}