+++ src/Rdd.Web/Querying/WebPageParser.cs
@@ -16,9 +16,14 @@ namespace Rdd.Web.Querying
         {
             _rddOptions = rddOptions;
         }
-        
+
         public Page Parse(HttpRequest request)
         {
+            if (!HttpMethods.IsGet(request.Method))
+            {
+                return Page.Unlimited;
+            }
+
             if (!request.Query.TryGetValue(Reserved.Paging, out var pageValue) || StringValues.IsNullOrEmpty(pageValue))
             {
                 return _rddOptions.Value.DefaultPage;
@@ -26,7 +31,7 @@ namespace Rdd.Web.Querying
 
             var input = pageValue.ToString();
 
-            if (input == "1") //...&paging=1 <=> &paging=0,100
+            if (input == "1") //...&paging=1 <=> &paging=0,10
             {
                 return _rddOptions.Value.DefaultPage;
             }
@@ -36,21 +41,21 @@ namespace Rdd.Web.Querying
 
                 if (elements.Length == 2)
                 {
-                    if (!Int32.TryParse(elements[0], out int offset))
+                    if (!int.TryParse(elements[0], out int offset))
                     {
-                        throw new BadRequestException(String.Format("Offset value {0} not in correct format", elements[0]));
+                        throw new BadRequestException(string.Format("Offset value {0} not in correct format", elements[0]));
                     }
 
-                    if (!Int32.TryParse(elements[1], out int limit))
+                    if (!int.TryParse(elements[1], out int limit))
                     {
-                        throw new BadRequestException(String.Format("Limit value {0} not in correct format", elements[1]));
+                        throw new BadRequestException(string.Format("Limit value {0} not in correct format", elements[1]));
                     }
 
                     return new Page(offset, limit, _rddOptions.Value.PagingMaximumLimit);
                 }
                 else
                 {
-                    throw new BadRequestException(String.Format("{0} does not respect limit=start,count format", input));
+                    throw new BadRequestException(string.Format("{0} does not respect limit=start,count format", input));
                 }
             }
         }