+++ src/Rdd.Web/Serialization/UrlProviders/IPluralizationService.cs
@@ -1,7 +1,0 @@
-﻿namespace Rdd.Web.Serialization.UrlProviders
-{
-    public interface IPluralizationService
-    {
-        string GetPlural(string name);
-    }
-}