+++ test/Rdd.Web.Tests/Serialization/UrlProviderTests.cs
@@ -1,7 +1,17 @@
-﻿using System;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc.Abstractions;
+using Microsoft.AspNetCore.Mvc.Controllers;
+using Microsoft.AspNetCore.Mvc.Infrastructure;
+using Microsoft.Extensions.DependencyInjection;
+using Moq;
+using Rdd.Web.Controllers;
 using Rdd.Web.Serialization.UrlProviders;
 using Rdd.Web.Tests.Models;
+using Rdd.Web.Tests.ServerMock;
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Reflection;
 using Xunit;
 
 namespace Rdd.Web.Tests.Serialization
@@ -9,6 +19,7 @@ namespace Rdd.Web.Tests.Serialization
     public class UrlProviderTests
     {
         private HttpContextAccessor httpContextAccessor;
+        private IActionDescriptorCollectionProvider actionDescriptorCollectionProvider;
 
         public UrlProviderTests()
         {
@@ -18,7 +29,24 @@ namespace Rdd.Web.Tests.Serialization
             };
             httpContextAccessor.HttpContext.Request.Scheme = "https";
             httpContextAccessor.HttpContext.Request.Host = new HostString("mon.domain.com");
+
+            var mock = new Mock<IActionDescriptorCollectionProvider>();
+            mock.Setup(a => a.ActionDescriptors).Returns(new ActionDescriptorCollection(new List<ActionDescriptor>
+            {
+                new ControllerActionDescriptor
+                {
+                    ActionName = "GetByIdAsync",
+                    ControllerTypeInfo = typeof(ReadOnlyWebController<User, int>).GetTypeInfo(),
+                    AttributeRouteInfo = new Microsoft.AspNetCore.Mvc.Routing.AttributeRouteInfo
+                    {
+                        Template = "api/users/{id}"
+                    }
+                }
+            }, 1));
+
+            actionDescriptorCollectionProvider = mock.Object;
         }
+
         [Fact]
         public void GetEntityUrl_should_return_valid_url()
         {
@@ -27,7 +55,7 @@ namespace Rdd.Web.Tests.Serialization
                 Id = 1
             };
 
-            var urlProvider = new UrlProvider(new PluralizationService(new Inflector.Inflector(new System.Globalization.CultureInfo("en-US"))), httpContextAccessor);
+            var urlProvider = new UrlProvider(actionDescriptorCollectionProvider, httpContextAccessor);
             var result = urlProvider.GetEntityApiUri(entity);
 
             Assert.Equal("https://mon.domain.com/api/users/1", result.ToString());
@@ -38,10 +66,10 @@ namespace Rdd.Web.Tests.Serialization
         {
             var entity = new UserTest();
 
-            var urlProvider = new UrlProvider(new PluralizationService(new Inflector.Inflector(new System.Globalization.CultureInfo("en-US"))), httpContextAccessor);
+            var urlProvider = new UrlProvider(actionDescriptorCollectionProvider, httpContextAccessor);
             var result = urlProvider.GetEntityApiUri(entity);
 
-            Assert.Equal("https://mon.domain.com/api/usertests/10", result.ToString());
+            Assert.Equal("https://mon.domain.com/api/users/10", result.ToString());
         }
 
         [Fact]
@@ -49,10 +77,32 @@ namespace Rdd.Web.Tests.Serialization
         {
             var entity = new UserTest();
 
-            var urlProvider = new UserTestUrlProvider(httpContextAccessor);
+            var urlProvider = new UserTestUrlProvider(actionDescriptorCollectionProvider, httpContextAccessor);
             var result = urlProvider.GetEntityApiUri(entity);
 
-            Assert.Equal("https://mon.domain.com/api/users/10", result.ToString());
+            Assert.Equal("https://mon.domain.com/specific/url/10", result.ToString());
+        }
+
+        [Fact]
+        public void UnknownObject_get_no_url()
+        {
+            var entity = new Department();
+
+            var urlProvider = new UserTestUrlProvider(actionDescriptorCollectionProvider, httpContextAccessor);
+            var result = urlProvider.GetEntityApiUri(entity);
+
+            Assert.Null(result);
+        }
+
+        [Fact]
+        public void RealUrls()
+        {
+            var host = Startup.BuildWebHost(null).Build();
+            var urlProvider = host.Services.GetRequiredService<IUrlProvider>();
+
+            var entity = new ExchangeRate { Id = 123 };
+            var result = urlProvider.GetEntityApiUri(entity);
+            Assert.Equal("https://mon.domain.com/ExchangeRates/123", result.ToString());
         }
 
         private class UserTest : User
@@ -65,17 +115,16 @@ namespace Rdd.Web.Tests.Serialization
 
         private class UserTestUrlProvider : UrlProvider
         {
-            public UserTestUrlProvider(IHttpContextAccessor httpContextAccessor)
-                : base(new PluralizationService(new Inflector.Inflector(new System.Globalization.CultureInfo("en-US"))), httpContextAccessor) { }
-
-            public override string GetApiControllerName(Type workingType)
+            public UserTestUrlProvider(IActionDescriptorCollectionProvider provider, IHttpContextAccessor httpContextAccessor)
+                : base(provider, httpContextAccessor)
             {
-                if (workingType == typeof(UserTest))
-                {
-                    return base.GetApiControllerName(typeof(User));
-                }
+            }
 
-                return base.GetApiControllerName(workingType);
+            protected override Dictionary<Type, string> CompileUrls()
+            {
+                var result = base.CompileUrls();
+                result[typeof(UserTest)] = "specific/url/{0}";
+                return result;
             }
         }
     }